Gw-55: Refactor out the home landing image into a shared component
