+++ src/pages/home/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import About from "./components/About";
 import { MainContainer } from "../../shared/layout/MainContainer";
-import { LandingImage } from "./components/LandingImage";
+import { HomeLandingImage } from "./components/HomeLandingImage";
 import Activities from "./components/activities/Activities";
 import { Events } from "./components/events/Events";
 import Partners from "./components/partners/Partners";
@@ -12,7 +12,7 @@ import { ScrollButtonContainer } from "./components/ScrollButton";
 export const Home = () => {
   return (
     <Box className="bg-[#FFFDF5]">
-      <LandingImage />
+      <HomeLandingImage />
       <MainContainer>
         <About />
         <Activities />
